requirements.txt
# ======================================================================
# Local Voice Assistant — Python Dependencies
# ======================================================================
# Install all:  pip install -r requirements.txt
# ======================================================================

# ── Core: Pipecat Voice Pipeline ──────────────────────────────────────
# Includes: local audio transport, Silero VAD, Whisper STT
pipecat-ai[local,silero,whisper]>=0.0.105
pyaudio>=0.2.14

# ── LLM: Ollama (local models) ────────────────────────────────────────
ollama>=0.4.4

# ── STT: Faster Whisper ───────────────────────────────────────────────
faster-whisper>=1.2.1

# ── TTS: Piper (offline neural TTS) ──────────────────────────────────
piper-tts>=1.2.0

# ── Audio playback ────────────────────────────────────────────────────
sounddevice==0.4.7
soundfile==0.12.1
numpy>=1.24.3

# ── AI/ML: PyTorch (needed for Whisper CUDA + Silero VAD) ────────────
torch>=2.0.0

# ── RAG: Embeddings + Vector Search ──────────────────────────────────
sentence-transformers>=2.2.2
scikit-learn>=1.3.0

# ── PDF Ingestion (upload_pdf.py) ─────────────────────────────────────
PyPDF2>=3.0.1

# ── Utilities ─────────────────────────────────────────────────────────
colorama>=0.4.6

# ── Optional: Wake Word Detection ─────────────────────────────────────
# Uncomment if you want "Hey Jarvis" wake word support
# openwakeword>=0.6.0

# ── Optional: Alternative TTS Engines ─────────────────────────────────
# Uncomment whichever engine you want to use:
# edge-tts>=6.1.9        # Microsoft Edge voices (needs internet)
# kokoro>=0.9.4          # High-quality neural TTS (needs ~1GB model download)
